requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2025.7.14
charset-normalizer==2.1.1
Django==3.2.25
django-filter==23.5
djangorestframework==3.15.1
idna==3.10
parameterized==0.9.0
pycodestyle==2.10.0
pytz==2025.2
requests==2.28.1
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==1.26.16
